medical-agent/requirements.txt
# Core dependencies
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
mysql-connector-python>=8.2.0
pandas>=2.0.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0

# MCP and A2A dependencies
pydantic>=2.5.0
httpx>=0.25.0

# Optional for CLI/web interface
typer>=0.9.0
streamlit>=1.28.0
